requirements.txt
html2markdown
tldextract
requests
pdfminer.six
